+++ test/rendering.test.js
@@ -179,8 +179,8 @@ describe('Rendering', () => {
       await testRender('variants-no-options', 'html', 200);
     });
 
-    it('renders no specifications, no name', async () => {
-      await testRender('no-specifications-no-name', 'html', 200);
+    it('renders no product name', async () => {
+      await testRender('no-product-name', 'html', 200);
     });
 
     it('renders no option uid', async () => {
@@ -213,6 +213,10 @@ describe('Rendering', () => {
 
     it('uses jsonld override from product instead of generating', async () => {
       await testRender('jsonld-override', 'html', 200);
+    });
+
+    it('renders aggregateRating nested on the Product node', async () => {
+      await testRender('aggregate-rating', 'html', 200);
     });
   });
 });